+++ app/Http/Controllers/UserController.php
@@ -66,7 +66,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        //
+        return view('user.show')->with(['user' => $user]);
     }
 
     /**